Added compose to DragNodesComposition.
